requirements.txt
django==5.2.1
asgiref==3.8.1
sqlparse==0.5.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dateutil==2.8.2
